refactor(stack): centralize managed workspace metadata
